# Conflicts:
#	README.md

+++ loadAndPushImage.go
@@ -9,28 +9,74 @@ import (
 	"strings"
 	"os/exec"
 	"syscall"
+	"flag"
 )
 
+var (
+	reg     string
+	pro     string
+	r       bool
+	push    bool
+	save    bool
+	h       bool
+	v       bool
+	VERSION string
+	ARCH    string
+)
+
+func init() {
+	flag.StringVar(&reg, "registry", "ygt", "镜像仓库名")
+	flag.StringVar(&pro, "project", "", "镜像仓库项目名")
+	flag.BoolVar(&r, "recursive", false, "是否递归子目录")
+	flag.BoolVar(&push, "push", true, "是否推送到镜像仓库")
+	flag.BoolVar(&save, "save", false, "是否保存修改后的镜像")
+	flag.BoolVar(&h, "help", false, "帮助信息")
+	flag.BoolVar(&v, "version", false, "版本信息")
+	flag.Usage = usage
+}
+
+func usage() {
+	fmt.Print(`遍历当前文件夹及其子文件夹,上传,推送,保存镜像
+  --registry string
+        required | 镜像仓库名 (default "ygt")
+  --project string
+        required | 镜像仓库项目名
+  --recursive
+        是否递归子目录
+  --push
+        是否推送到镜像仓库 (default true)
+  --save
+        是否保存修改后的镜像
+  --version
+        版本信息
+  --help
+        帮助信息
+`)
+}
+
 func main() {
-	if len(os.Args) == 1 || len(os.Args) == 5 {
-		fmt.Println("缺少参数，请参考 --help")
+	flag.Parse()
+	if h {
+		usage()
+		os.Exit(1)
+	}
+
+	if v {
+		fmt.Printf("VERSION: %s\n", VERSION)
+		fmt.Printf("ARCH: %s\n", ARCH)
+		os.Exit(1)
+	}
+
+	if len(os.Args) == 1 {
+		usage()
 		os.Exit(0)
 	}
 
-	if strings.Contains(os.Args[1], "--help") {
-		fmt.Println("usage:./xxxx args1 args2 true \n " +
-			"args1: \t required 仓库地址 \n " +
-			"args2: \t required 项目地址 没有请填写nil \n " +
-			"true | false: \t required  是否递归当前文件夹下的子文件夹 \n " +
-			"push: \t required  是否push \n " +
-			"save: \t required  是否save \n")
+	if reg == "" || pro == "" {
+		fmt.Println("不符合官方推荐命名,请手动执行...")
 		os.Exit(0)
 	}
 
-	if os.Args[1] == "nil" && os.Args[2] == "nil" {
-		fmt.Println("不符合官方推荐命名，请手动执行...")
-		os.Exit(0)
-	}
 	path := getCurrentDir()
 	fileList := list.New()
 	i := readDir(path, fileList)
@@ -54,20 +100,24 @@ func main() {
 				} else {
 					newImageName = getNewImageName(oldImageName)
 				}
+				oldImageName = strings.TrimSpace(oldImageName)
+				newImageName = strings.TrimSpace(newImageName)
 				if k == 0 {
-					if !strings.EqualFold(oldImageName, newImageName) {
+					if !(oldImageName == newImageName) {
 						tagImage(oldImageName, newImageName)
-						deleteImage(oldImageName)
+						deleteImage(oldImageName, newImageName)
 					}
-					if os.Args[4] == "push" {
+
+					if push {
 						pushImage(newImageName)
 					}
-					if os.Args[5] == "save" {
+
+					if save {
 						saveImage(newImageName)
 					}
 				}
 				if k > 0 {
-					deleteImage(oldImageName)
+					deleteImage(oldImageName, newImageName)
 				}
 			}
 		} else {
@@ -78,28 +128,36 @@ func main() {
 				imageName := split[len(split)-1]
 				//newImageName := os.Args[1] + "/" + os.Args[2] + "/" + imageName
 				newImageName := getNewImageName(imageName)
-				if !strings.EqualFold(oldImageName, newImageName) {
+				oldImageName = strings.TrimSpace(oldImageName)
+				newImageName = strings.TrimSpace(newImageName)
+				if !(oldImageName == newImageName) {
 					tagImage(oldImageName, newImageName)
-					deleteImage(oldImageName)
+					deleteImage(oldImageName, newImageName)
 				}
-				if os.Args[4] == "push" {
+
+				if push {
 					pushImage(newImageName)
 				}
-				if os.Args[5] == "save" {
+
+				if save {
 					saveImage(newImageName)
 				}
 			} else {
 				oldImageName := strings.Replace(result, "Loaded image:", "", -1)
 				//newImageName := os.Args[1] + "/" + os.Args[2] + "/" + oldImageName
 				newImageName := getNewImageName(oldImageName)
-				if !strings.EqualFold(oldImageName, newImageName) {
+				oldImageName = strings.TrimSpace(oldImageName)
+				newImageName = strings.TrimSpace(newImageName)
+				if !(oldImageName == newImageName) {
 					tagImage(oldImageName, newImageName)
-					deleteImage(oldImageName)
+					deleteImage(oldImageName, newImageName)
 				}
-				if os.Args[4] == "push" {
+
+				if push {
 					pushImage(newImageName)
 				}
-				if os.Args[5] == "save" {
+
+				if save {
 					saveImage(newImageName)
 				}
 			}
@@ -109,12 +167,12 @@ func main() {
 
 func getNewImageName(imageName string) string {
 	var newImageName string
-	if os.Args[2] == "nil" {
-		newImageName = os.Args[1] + "/" + imageName
+	if pro == "" {
+		newImageName = reg + "/" + imageName
 	} else {
-		newImageName = os.Args[1] + "/" + os.Args[2] + "/" + imageName
+		newImageName = reg + "/" + pro + "/" + imageName
 	}
-	return newImageName
+	return strings.TrimSpace(newImageName)
 }
 
 func tagImage(oldImageName string, newImageName string) {
@@ -125,7 +183,7 @@ func tagImage(oldImageName string, newImageName string) {
 }
 
 func pushImage(newImageName string) {
-	pushShell := strings.Join([]string{" && docker push ", newImageName}, " ")
+	pushShell := strings.Join([]string{"docker push ", newImageName}, " ")
 	fmt.Print("正在执行...")
 	fmt.Println(pushShell)
 	execCommand(pushShell)
@@ -141,7 +199,10 @@ func saveImage(newImageName string) {
 	execCommand(saveShell)
 }
 
-func deleteImage(imageName string) {
+func deleteImage(imageName, newImageName string) {
+	if imageName == newImageName {
+		return
+	}
 	rmShell := strings.Join([]string{"docker rmi ", imageName}, " ")
 	fmt.Print("正在执行...")
 	fmt.Println(rmShell)
@@ -164,7 +225,7 @@ func readDir(path string, fileList *list.List) *list.List {
 	for _, file := range infos {
 		fileName := file.Name()
 		if file.IsDir() {
-			if os.Args[3] == "true" {
+			if r {
 				readDir(path+"/"+fileName, fileList)
 			}
 		} else {
